fixed boss remarks 2
